Added featurehub, needs testing

+++ BPT-System/MeasurementService/Controllers/MeasurementController.cs
@@ -1,3 +1,4 @@
+using FeatureHubSDK;
 using Microsoft.AspNetCore.Mvc;
 using Models;
 using MeasurementService.Repositories;
@@ -6,17 +7,23 @@ namespace MeasurementService.Controllers;
 
 [ApiController]
 [Route("[controller]")]
-public class MeasurementController(IMeasurementRepository measurementRepository) : ControllerBase
+public class MeasurementController(IMeasurementRepository measurementRepository, IClientContext fh) : ControllerBase
 {
     [HttpGet("GetMeasurementById")]
     public async Task<ActionResult<Measurement>> GetMeasurementById([FromQuery] int id, CancellationToken ct) 
     {
+        if (fh["MeasurementServiceOff"].IsEnabled) {
+            return NoContent();
+        }
         return await measurementRepository.GetByIdAsync(id, ct);
     }
 
     [HttpPost("AddMeasurement")]
     public async Task<IActionResult> AddMeasurement([FromBody] Measurement measurement, CancellationToken ct)
     {
+        if (fh["MeasurementServiceOff"].IsEnabled) {
+            return NoContent();
+        }
         await measurementRepository.AddAsync(measurement, ct);
         return Ok();
     }
@@ -24,6 +31,9 @@ public class MeasurementController(IMeasurementRepository measurementRepository)
     [HttpPut("UpdateMeasurement")]
     public async Task<IActionResult> UpdateMeasurement([FromBody] Measurement measurement, CancellationToken ct)
     {
+        if (fh["MeasurementServiceOff"].IsEnabled) {
+            return NoContent();
+        }
         await measurementRepository.UpdateAsync(measurement, ct);
         return Ok();
     }
@@ -31,6 +41,9 @@ public class MeasurementController(IMeasurementRepository measurementRepository)
     [HttpDelete("DeleteMeasurement")]
     public async Task<IActionResult> DeleteMeasurement([FromBody] Measurement measurement, CancellationToken ct)
     {
+        if (fh["MeasurementServiceOff"].IsEnabled) {
+            return NoContent();
+        }
         await measurementRepository.DeleteAsync(measurement, ct);
         return Ok();
     }
@@ -38,6 +51,9 @@ public class MeasurementController(IMeasurementRepository measurementRepository)
     [HttpDelete("DeleteMeasurementById")]
     public async Task<IActionResult> DeleteMeasurementById([FromQuery] int id, CancellationToken ct)
     {
+        if (fh["MeasurementServiceOff"].IsEnabled) {
+            return NoContent();
+        }
         await measurementRepository.DeleteByIdAsync(id, ct);
         return Ok();
     }
