+++ BPT-System/PatientService/Controllers/PatientController.cs
@@ -1,3 +1,4 @@
+using FeatureHubSDK;
 using Microsoft.AspNetCore.Mvc;
 using Models;
 using PatientService.Repositories;
@@ -6,18 +7,26 @@ namespace PatientService.Controllers;
 
 [ApiController]
 [Route("[controller]")]
-public class PatientController(IPatientRepository patientRepository) : ControllerBase
+public class PatientController(IPatientRepository patientRepository, IClientContext fh) : ControllerBase
 {
 
     [HttpGet("GetPatientBySsn")]
     public async Task<ActionResult<Patient>> GetPatientBySsn([FromQuery] string ssn, CancellationToken ct)
     {
+        if (fh["PatientServiceOff"].IsEnabled) {
+            return NoContent();
+        }
+
         return await patientRepository.GetBySsnAsync(ssn, ct);
     }
     
     [HttpGet("GetPatientMeasurementsBySsn")]
     public async Task<ActionResult<MeasurementsOfPatientDto>> GetPatientMeasurementsBySsn([FromQuery] string ssn, CancellationToken ct)
     {
+        if (fh["PatientServiceOff"].IsEnabled) {
+            return NoContent();
+        }
+        
         return await patientRepository.GetMeasurementsOfPatientAsync(ssn, ct);
     }
  
@@ -25,6 +34,10 @@ public class PatientController(IPatientRepository patientRepository) : Controlle
     [HttpPost("AddPatient")]
     public async Task<IActionResult> AddPatient([FromBody] Patient patient, CancellationToken ct)
     {
+        if (fh["PatientServiceOff"].IsEnabled) {
+            return NoContent();
+        }
+        
         await patientRepository.AddAsync(patient, ct);
         return Ok();
     }
@@ -32,6 +45,10 @@ public class PatientController(IPatientRepository patientRepository) : Controlle
     [HttpPut("UpdatePatient")]
     public async Task<IActionResult> UpdatePatient([FromBody] Patient patient, CancellationToken ct)
     {
+        if (fh["PatientServiceOff"].IsEnabled) {
+            return NoContent();
+        }
+        
         await patientRepository.UpdateAsync(patient, ct);
         return Ok();
     }
@@ -39,6 +56,10 @@ public class PatientController(IPatientRepository patientRepository) : Controlle
     [HttpDelete("DeletePatient")]
     public async Task<IActionResult> DeletePatient([FromBody] Patient patient, CancellationToken ct) 
     {
+        if (fh["PatientServiceOff"].IsEnabled) {
+            return NoContent();
+        }
+        
         await patientRepository.DeleteAsync(patient, ct);
         return Ok();
     }
